Use tested incognito browser session for short-lived MCP tools
This is a follow-up to PR #599, see that PR for the details.

Note that those short-lived MCP tools are for NPR, CNN, etc.
