+++ getgather/mcp/dpage.py
@@ -21,6 +21,7 @@ from getgather.distill import (
     capture_page_artifacts,
     convert,
     distill,
+    get_incognito_browser_profile,
     get_selector,
     load_distillation_patterns,
     report_distill_error,
@@ -321,42 +322,6 @@ def is_local_address(host: str) -> bool:
         return ip.is_loopback
     except ValueError:
         return hostname in ("localhost", "127.0.0.1")
-
-
-async def get_incognito_browser_profile(signin_id: str | None) -> BrowserProfile:
-    """Get or create an incognito browser profile."""
-    if signin_id is not None:
-        if signin_id in incognito_browser_profiles:
-            return incognito_browser_profiles[signin_id]
-        else:
-            raise ValueError(f"Browser profile for signin {signin_id} not found")
-
-    MAX_ATTEMPTS = 3
-    CHECK_URL = "https://ip.fly.dev/all"
-    CHECK_TIMEOUT = 10  # seconds
-    for attempt in range(1, MAX_ATTEMPTS + 1):
-        logger.info(f"Creating incognito browser profile (attempt {attempt}/{MAX_ATTEMPTS})...")
-        fresh_profile = BrowserProfile()
-        fresh_session = BrowserSession.get(fresh_profile)
-
-        try:
-            await fresh_session.start(debug_url=None)
-            check_page = await fresh_session.new_page()
-            logger.error(f"Validating incognito browser profile at {CHECK_URL}...")
-            await check_page.goto(CHECK_URL, timeout=CHECK_TIMEOUT * 1000)
-            logger.info(f"Incognito browser profile validated on attempt {attempt}")
-            return fresh_profile
-
-        except Exception as e:
-            logger.warning(f"Incognito browser profile validation failed on attempt {attempt}: {e}")
-            if attempt < MAX_ATTEMPTS:
-                try:
-                    await fresh_session.stop()
-                except Exception:
-                    pass
-
-    logger.error(f"Failed to get browser profile after {MAX_ATTEMPTS} attempts!")
-    raise RuntimeError(f"Failed to get browser profile after {MAX_ATTEMPTS} attempts!")
 
 
 async def dpage_mcp_tool(initial_url: str, result_key: str, timeout: int = 2) -> dict[str, Any]:
